Data importing

We have a number of ways to import data, besides using any of the additional packages in julia to read files.

GeophysicalModelGenerator.Screenshot_To_GeoData — Function

Screenshot_To_GeoData(filename::String, Corner_LowerLeft, Corner_UpperRight; Corner_LowerRight=nothing, Corner_UpperLeft=nothing)

Take a screenshot of Georeferenced image (either a lat/lon map at a given depth or a profile) and converts it to a GeoData struct, which can be saved to Paraview

The lower left and upper right coordinates of the image need to be specified in tuples of (lon,lat,depth), where depth is negative in the Earth.

The lower right and upper left corners can be specified optionally (to take non-orthogonal images into account). If they are not specified, the image is considered orthogonal and the corners are computed from the other two.

GeophysicalModelGenerator.Screenshot_To_CartData — Function

Data = Screenshot_To_CartData(filename::String, Corner_LowerLeft, Corner_UpperRight; Corner_LowerRight=nothing, Corner_UpperLeft=nothing)

Does the same as Screenshot_To_GeoData, but returns a Cartesian data structure