pTatin

We provide a few routines that allow exporting GMG input data sets to a format that can be read by the 3D geodynamic modelling software pTatin, which is an open-source Cartesian code to perform crustal and lithospheric-scale simulations.

The Q1Data format can be saved directly to pTatin.

GeophysicalModelGenerator.write_pTatin_mesh — Function

write_pTatin_mesh(fe_mesh::FEData; out_file="md.bin", connectivity_zero_based=true)

Write a binary file with the mesh information for pTatin

write_pTatin_mesh(q1_mesh::Q1Data; out_file="md.bin", connectivity_zero_based=true)

Write a binary file with the mesh information for pTatin

GeophysicalModelGenerator.swap_yz_dims — Function

fe_swap = swap_yz_dims(fe_data::FEData)

This swaps the y and z dimensions of the FEData object, which is useful for pTatin as it uses y for what is z in GMG.