Data importing

We have a number of ways to import data, besides using any of the additional packages in julia to read files.

GeophysicalModelGenerator.screenshot_to_GeoData — Function

screenshot_to_GeoData(filename::String, Corner_LowerLeft, Corner_UpperRight; Corner_LowerRight=nothing, Corner_UpperLeft=nothing, Cartesian=false, UTM=false, UTMzone, isnorth=true, fieldname::Symbol=:colors)

Take a screenshot of Georeferenced image either a lat/lon, x,y (if Cartesian=true) or in UTM coordinates (if UTM=true) at a given depth or along profile and converts it to a GeoData, CartData or UTMData struct, which can be saved to Paraview

The lower left and upper right coordinates of the image need to be specified in tuples of (lon,lat,depth) or (UTM_ew, UTM_ns, depth), where depth is negative inside the Earth (and in km).

The lower right and upper left corners can be specified optionally (to take non-orthogonal images into account). If they are not specified, the image is considered orthogonal and the corners are computed from the other two.

Note: if your data is in UTM coordinates you also need to provide the UTMzone and whether we are on the northern hemisphere or not (isnorth).

GeophysicalModelGenerator.screenshot_to_CartData — Function

Data = screenshot_to_CartData(filename::String, Corner_LowerLeft, Corner_UpperRight; Corner_LowerRight=nothing, Corner_UpperLeft=nothing)

Does the same as screenshot_to_GeoData, but returns a CartData structure

GeophysicalModelGenerator.screenshot_to_UTMData — Function

Data = screenshot_to_UTMData(filename::String, Corner_LowerLeft, Corner_UpperRight; Corner_LowerRight=nothing, Corner_UpperLeft=nothing, UTMzone::Int64=nothing, isnorth::Bool=true, fieldname=:colors)

Does the same as screenshot_to_GeoData, but returns for UTM data Note that you have to specify the UTMzone and isnorth

GeophysicalModelGenerator.import_topo — Function

    import_topo

Optional routine that imports topography. It requires you to load GMT

GeophysicalModelGenerator.import_GeoTIFF — Function

    import_GeoTIFF

Optional routine that imports GeoTIFF images. It requires you to load GMT

GeophysicalModelGenerator.tomo_2_GeoData — Function

Read_TomoData(filename::String)

Reads a seismic tomography dataset from a NetCDF file as a GeoData object. The keyword argument vel_type::String allows you to specify the type of velocity data to extract (default is "vs" for shear wave velocity). The function assumes that the NetCDF file contains variables for depth, longitude, latitude, and the specified velocity type.

tomodata = tomo2GeoData("path/to/tomo_data.nc")